+++ libs/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphql-auto-queries/core
 
+## 0.0.6
+
+### Patch Changes
+
+- Fix deep nested dataloaders
+
+## 0.0.5
+
+### Patch Changes
+
+- Fix bug on null field resolver
+
 ## 0.0.4
 
 ### Patch Changes