+++ libs/sql-converter/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @graphql-auto-queries/sql-converter
 
+## 0.0.6
+
+### Patch Changes
+
+- Fix deep nested dataloaders
+- Updated dependencies
+  - @graphql-auto-queries/core@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphql-auto-queries/core@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes